Refactor into workspace with library and CLI crates.
