+++ Cargo.toml
@@ -1,8 +1,3 @@
-[package]
-name = "cards-rust"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-printpdf = { version = "0.8", features = ["png", "jpeg"] }
-clap = { version = "4.5", features = ["derive"] }
+[workspace]
+members = ["cards-core", "cards-cli"]
+resolver = "2"
