+++ cards-cli/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "cards-cli"
+version = "0.1.0"
+edition = "2021"
+
+[[bin]]
+name = "cards"
+path = "src/main.rs"
+
+[dependencies]
+cards-core = { path = "../cards-core" }
+clap = { version = "4.5", features = ["derive"] }
+env_logger = "0.11"
+log = "0.4"
